app/requirements.txt
amqp==5.0.9
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.11
django-celery-beat==2.2.1
django-import-export==2.7.1
django-leaflet==0.28.2
django-timezone-field==4.2.3
gdal==3.2.2
gunicorn==20.1.0
idna==3.3
kombu==5.2.3
prompt-toolkit==3.0.26
psycopg2-binary==2.9.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5

